Make changes to DailaySalesUI & changed my UI to a Panel
