+++ src/itemmanagement/ItemManagement.java
@@ -21,7 +21,7 @@ import java.awt.event.MouseEvent;
 public class ItemManagement implements ItemOperations {
     //----- Basically these are passed from the UI.java -----//
     private JTable itemTable;
-    private JTextField itemName_textbox, itemCode_textbox, category_textbox, unitPrice_textbox, retailPrice_textbox, description_textbox;
+    private JTextField itemName_textbox, itemCode_textbox, unitPrice_textbox, retailPrice_textbox;
     private JComboBox supplierID_comboBox, supplierName_comboBox;
     private TableRowSorter<DefaultTableModel> sorter;
 
@@ -31,19 +31,15 @@ public class ItemManagement implements ItemOperations {
                                             JTextField code, 
                                             JComboBox supID, 
                                             JComboBox supName, 
-                                            JTextField category, 
                                             JTextField price, 
-                                            JTextField rprice,
-                                            JTextField desc) {
+                                            JTextField rprice ) {
         this.itemTable = table;
         this.itemName_textbox = name;
         this.itemCode_textbox = code;
         this.supplierID_comboBox = supID;
         this.supplierName_comboBox = supName;
-        this.category_textbox = category;
         this.unitPrice_textbox = price;
         this.retailPrice_textbox = rprice;
-        this.description_textbox = desc;
     }
     
 
@@ -55,13 +51,11 @@ public class ItemManagement implements ItemOperations {
         String code = itemCode_textbox.getText().trim();
         String supID = (String) supplierID_comboBox.getSelectedItem();
         String supName = (String) supplierName_comboBox.getSelectedItem();
-        String category = category_textbox.getText().trim();
         String price = unitPrice_textbox.getText().trim();
         String rprice = retailPrice_textbox.getText().trim();
-        String desc = description_textbox.getText().trim();
         
         //----- Check for errors before adding to the table -----//
-        if (name.isEmpty() || code.isEmpty()  || supID == null || supName == null || category.isEmpty() || price.isEmpty() || rprice.isEmpty() || desc.isEmpty()) {
+        if (name.isEmpty() || code.isEmpty()  || supID == null || supName == null ||  price.isEmpty() || rprice.isEmpty()) {
             JOptionPane.showMessageDialog(null, "Please fill up everything before adding to table");
             return;
         }
@@ -73,10 +67,8 @@ public class ItemManagement implements ItemOperations {
                 code.equalsIgnoreCase(itemTable.getValueAt(i, 2).toString()) &&  
                 supID.equals(itemTable.getValueAt(i, 3).toString()) &&          
                 supName.equals(itemTable.getValueAt(i, 4).toString()) &&        
-                category.equalsIgnoreCase(itemTable.getValueAt(i, 5).toString()) &&
-                price.equals(itemTable.getValueAt(i, 7).toString()) &&
-                rprice.equals(itemTable.getValueAt(i, 8).toString()) &&
-                desc.equalsIgnoreCase(itemTable.getValueAt(i, 9).toString())
+                price.equals(itemTable.getValueAt(i, 6).toString()) &&
+                rprice.equals(itemTable.getValueAt(i, 7).toString()) 
             ) {
             JOptionPane.showMessageDialog(null, "This item already exists in the table.");
             return;
@@ -86,7 +78,7 @@ public class ItemManagement implements ItemOperations {
         //----- Now add the data to the table -----//
         DefaultTableModel model = (DefaultTableModel) itemTable.getModel();
         int rowCount = model.getRowCount() + 1;
-        model.addRow(new Object[] {rowCount, name, code, supID, supName, category, "0", price, rprice, desc, false});
+        model.addRow(new Object[] {rowCount, name, code, supID, supName, "0", price, rprice, false});
          saveTableToFile();
         JOptionPane.showMessageDialog(null, "Item added");
         
@@ -95,10 +87,8 @@ public class ItemManagement implements ItemOperations {
         itemCode_textbox.setText("");
         supplierID_comboBox.setSelectedIndex(-1);
         supplierName_comboBox.setSelectedIndex(-1);
-        category_textbox.setText("");
         unitPrice_textbox.setText("");
         retailPrice_textbox.setText("");
-        description_textbox.setText("");
     }
     
     
@@ -113,13 +103,11 @@ public class ItemManagement implements ItemOperations {
             String code = itemCode_textbox.getText().trim();
             String supID = (String) supplierID_comboBox.getSelectedItem();
             String supName = (String) supplierName_comboBox.getSelectedItem();
-            String category = category_textbox.getText().trim();
             String price = unitPrice_textbox.getText().trim();
             String rprice = retailPrice_textbox.getText().trim();
-            String desc = description_textbox.getText().trim();
 
             //----- Check if any fields are empty -----//
-            if (name.isEmpty() || code.isEmpty() || supID == null || supName == null || category.isEmpty() || price.isEmpty() || rprice.isEmpty() || desc.isEmpty()) {
+            if (name.isEmpty() || code.isEmpty() || supID == null || supName == null || price.isEmpty() || rprice.isEmpty()) {
                 JOptionPane.showMessageDialog(null, "Please fill up everything before updating.");
                 return;
             }
@@ -132,10 +120,8 @@ public class ItemManagement implements ItemOperations {
                     code.equalsIgnoreCase(itemTable.getValueAt(i, 2).toString()) &&
                     supID.equals(itemTable.getValueAt(i, 3).toString()) &&
                     supName.equals(itemTable.getValueAt(i, 4).toString()) &&
-                    category.equalsIgnoreCase(itemTable.getValueAt(i, 5).toString()) &&
-                    price.equals(itemTable.getValueAt(i, 7).toString()) &&
-                    rprice.equals(itemTable.getValueAt(i, 8).toString()) &&
-                    desc.equalsIgnoreCase(itemTable.getValueAt(i, 9).toString())
+                    price.equals(itemTable.getValueAt(i, 6).toString()) &&
+                    rprice.equals(itemTable.getValueAt(i, 7).toString()) 
                 ) {
                     JOptionPane.showMessageDialog(null, "This item already exists in the table.");
                     return;
@@ -147,10 +133,8 @@ public class ItemManagement implements ItemOperations {
             model.setValueAt(code, selectedRow, 2);
             model.setValueAt(supID, selectedRow, 3);
             model.setValueAt(supName, selectedRow, 4);
-            model.setValueAt(category, selectedRow, 5);
-            model.setValueAt(price, selectedRow, 7);
-            model.setValueAt(rprice, selectedRow, 8);
-            model.setValueAt(desc, selectedRow, 9);
+            model.setValueAt(price, selectedRow, 6);
+            model.setValueAt(rprice, selectedRow, 7);
 
             //----- Save to items.txt -----//
             saveTableToFile();
@@ -160,10 +144,8 @@ public class ItemManagement implements ItemOperations {
             itemCode_textbox.setText("");
             supplierID_comboBox.setSelectedIndex(-1);
             supplierName_comboBox.setSelectedIndex(-1);
-            category_textbox.setText("");
             unitPrice_textbox.setText("");
             retailPrice_textbox.setText("");
-            description_textbox.setText("");
 
         } else {
             JOptionPane.showMessageDialog(null, "Please select a row to update");
@@ -205,10 +187,8 @@ public class ItemManagement implements ItemOperations {
                         itemCode_textbox.setText(itemTable.getValueAt(selectedRow, 2).toString());
                         supplierID_comboBox.setSelectedItem(itemTable.getValueAt(selectedRow, 3).toString());
                         supplierName_comboBox.setSelectedItem(itemTable.getValueAt(selectedRow, 4).toString());
-                        category_textbox.setText(itemTable.getValueAt(selectedRow, 5).toString());
                         unitPrice_textbox.setText(itemTable.getValueAt(selectedRow, 7).toString());
                         retailPrice_textbox.setText(itemTable.getValueAt(selectedRow, 8).toString());
-                        description_textbox.setText(itemTable.getValueAt(selectedRow, 9).toString());
                     }
                 }
             }
@@ -223,10 +203,8 @@ public class ItemManagement implements ItemOperations {
         itemCode_textbox.setText("");
         supplierID_comboBox.setSelectedIndex(-1);
         supplierName_comboBox.setSelectedIndex(-1);
-        category_textbox.setText("");
         unitPrice_textbox.setText("");
         retailPrice_textbox.setText("");
-        description_textbox.setText("");
     }
     
     
@@ -305,7 +283,7 @@ public class ItemManagement implements ItemOperations {
             
             while ((line = reader.readLine()) != null) {
                 String[] data = line.split(",");
-                if (data.length == 10) { 
+                if (data.length == 8) { 
                     model.addRow(new Object[] {
                         no++,       //----- No. 
                         data[0],    
@@ -315,9 +293,7 @@ public class ItemManagement implements ItemOperations {
                         data[4],
                         data[5],
                         data[6],
-                        data[7],
-                        data[8],
-                        data[9]//----- Description
+                        data[7]     //----- DeliveryStatus
                     });
                 }
             }
