+++ src/itemmanagement/items.txt
@@ -1,1 +1,1 @@
-Dyson,DS0001,SUP0001,Nelson Tan,HOME,100,150.00,200.00,Vacuum!,false
+Dyson,DS0001,SUP0001,Nelson Tan,100,150.00,200.00,false
